desloppify: split mcp/server.rs into dispatch/audit_envelope/tool_name/tool_catalog
