+++ crates/rimap-server/src/mcp/server.rs
@@ -12,9 +12,8 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 use rimap_audit::AuditWriter;
-use rimap_audit::record::{Provenance, ResultSummary, ToolStatus};
-use rimap_audit::redact::{RedactionSalt, RedactionSchema, Redactor, hash_arguments, schemas};
-use rimap_core::account::{AccountId, DEFAULT_ACCOUNT_NAME};
+use rimap_audit::redact::{RedactionSalt, RedactionSchema, schemas};
+use rimap_core::account::AccountId;
 use rimap_core::tool::ToolName;
 use rmcp::RoleServer;
 use rmcp::handler::server::ServerHandler;
@@ -25,7 +24,10 @@ use rmcp::model::{
 };
 use rmcp::service::RequestContext;
 
-use crate::boot::registry::{AccountRegistry, AccountState};
+use crate::boot::registry::AccountRegistry;
+use crate::mcp::dispatch::{PostureContext, rimap_error_to_breaker_reason};
+use crate::mcp::tool_catalog::TOOL_DEFS;
+use crate::mcp::tool_name::{is_legacy_single_account, refine_tool_name, split_tool_name};
 
 /// Core MCP server. Owns every resource the handler methods need.
 pub struct ImapMcpServer {
@@ -275,875 +277,5 @@ impl ServerHandler for ImapMcpServer {
             result
         })
         .await
-    }
-}
-
-/// Posture context recorded in audit envelope headers.
-///
-/// Per-account dispatches use the account's effective posture; the
-/// infrastructure tools (`list_accounts`, `use_account`) bypass posture
-/// gating by design and record the dedicated `Infrastructure` variant so
-/// log readers can distinguish them from per-account dispatches.
-#[derive(Debug, Clone, Copy)]
-enum PostureContext {
-    Account(rimap_core::Posture),
-    Infrastructure,
-}
-
-impl PostureContext {
-    /// The per-account [`rimap_core::Posture`] this context represents, or
-    /// `None` for the infrastructure dispatch path. The audit writer maps
-    /// `None` to the `"infrastructure"` sentinel it records on disk.
-    fn posture(self) -> Option<rimap_core::Posture> {
-        match self {
-            Self::Account(p) => Some(p),
-            Self::Infrastructure => None,
-        }
-    }
-}
-
-/// Map a [`RimapError`] to the breaker's [`FailureReason`], or `None`
-/// when the error represents a user/agent/policy failure (which the
-/// breaker must ignore per its contract).
-fn rimap_error_to_breaker_reason(
-    err: &rimap_core::RimapError,
-) -> Option<rimap_authz::breaker::FailureReason> {
-    use rimap_authz::breaker::FailureReason;
-    use rimap_core::ErrorCode;
-    match err.code() {
-        ErrorCode::ConnectionLost => Some(FailureReason::ConnectionLost),
-        ErrorCode::Auth => Some(FailureReason::Auth),
-        ErrorCode::Timeout => Some(FailureReason::Timeout),
-        ErrorCode::ImapProtocol | ErrorCode::SmtpProtocol => Some(FailureReason::Protocol),
-        ErrorCode::Tls => Some(FailureReason::Tls),
-        ErrorCode::InvalidInput
-        | ErrorCode::PostureDenied
-        | ErrorCode::RateLimited
-        | ErrorCode::CircuitOpen
-        | ErrorCode::NotFound
-        | ErrorCode::AttachmentTooLarge
-        | ErrorCode::ProtectedFolder
-        | ErrorCode::ExpungeDenied
-        | ErrorCode::Config
-        | ErrorCode::Internal
-        | ErrorCode::NoAccount
-        | ErrorCode::UnknownAccount => None,
-    }
-}
-
-impl ImapMcpServer {
-    /// Wrap an inner dispatch `body` in the full audit envelope:
-    /// redact+hash args, emit `tool_start`, time the body, emit
-    /// `tool_end` with the status/error code derived from the body's
-    /// result. Returns the MCP-shaped `CallToolResult` or `ErrorData`.
-    async fn run_with_audit_envelope<F>(
-        &self,
-        tool: ToolName,
-        audit_account: Option<String>,
-        posture: PostureContext,
-        args: &serde_json::Map<String, serde_json::Value>,
-        body: F,
-    ) -> Result<CallToolResult, ErrorData>
-    where
-        F: std::future::Future<Output = Result<serde_json::Value, rimap_core::RimapError>>,
-    {
-        let args_value = serde_json::Value::Object(args.clone());
-        let redacted = self.redact_tool_args(tool, &args_value);
-        let hash = hash_arguments(&args_value);
-
-        let start_seq = self
-            .emit_tool_start(tool, audit_account.clone(), posture, redacted, hash)
-            .await?;
-        let start_time = std::time::Instant::now();
-
-        let result = body.await;
-
-        let duration_ms = start_time
-            .elapsed()
-            .as_millis()
-            .try_into()
-            .unwrap_or(u64::MAX);
-        let (status, error_code) = match &result {
-            Ok(_) => (ToolStatus::Ok, None),
-            Err(e) => (ToolStatus::Error, Some(e.code())),
-        };
-        self.emit_tool_end(
-            start_seq,
-            tool,
-            audit_account,
-            status,
-            error_code,
-            duration_ms,
-        )
-        .await;
-
-        match result {
-            Ok(value) => Ok(CallToolResult::structured(value)),
-            Err(e) => Err(crate::mcp::error::to_mcp_error(&e)),
-        }
-    }
-
-    /// Dispatch to the tool handler for `tool`. Each arm serializes its
-    /// typed response to `serde_json::Value` before returning, so the
-    /// audit envelope works with a single future type.
-    #[doc(hidden)]
-    pub async fn dispatch_tool(
-        &self,
-        account: &AccountState,
-        tool: ToolName,
-        args: &serde_json::Map<String, serde_json::Value>,
-    ) -> Result<serde_json::Value, rimap_core::RimapError> {
-        use crate::tools::admin::list_folders;
-        use crate::tools::compose::{create_draft, send_email};
-        use crate::tools::mailbox::{
-            delete_message, expunge, flags, folder_management, labels, move_message,
-        };
-        use crate::tools::retrieval::{
-            download_attachment, fetch_message, list_attachments, search,
-        };
-        let resp = match tool {
-            ToolName::ListFolders => ser(Box::pin(list_folders::handle(account)).await?)?,
-            ToolName::MarkRead => {
-                ser(Box::pin(flags::handle_mark_read(account, parse_args(args)?)).await?)?
-            }
-            ToolName::MarkUnread => {
-                ser(Box::pin(flags::handle_mark_unread(account, parse_args(args)?)).await?)?
-            }
-            ToolName::Flag => ser(Box::pin(flags::handle_flag(account, parse_args(args)?)).await?)?,
-            ToolName::Unflag => {
-                ser(Box::pin(flags::handle_unflag(account, parse_args(args)?)).await?)?
-            }
-            ToolName::MoveMessage => {
-                ser(Box::pin(move_message::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::Search | ToolName::SearchAdvanced => {
-                ser(Box::pin(search::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::FetchMessage | ToolName::FetchMessageHtml => {
-                ser(Box::pin(fetch_message::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::ListAttachments => {
-                ser(Box::pin(list_attachments::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::DownloadAttachment => {
-                let input = parse_args(args)?;
-                ser(Box::pin(download_attachment::handle(
-                    account,
-                    input,
-                    &self.download_dir,
-                ))
-                .await?)?
-            }
-            ToolName::CreateDraft => {
-                ser(Box::pin(create_draft::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::SendEmail => {
-                ser(Box::pin(send_email::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::DeleteMessage => {
-                ser(Box::pin(delete_message::handle(account, parse_args(args)?)).await?)?
-            }
-            ToolName::Expunge => ser(Box::pin(expunge::handle(account, parse_args(args)?)).await?)?,
-            ToolName::CreateFolder => ser(Box::pin(folder_management::handle_create_folder(
-                account,
-                parse_args(args)?,
-            ))
-            .await?)?,
-            ToolName::RenameFolder => ser(Box::pin(folder_management::handle_rename_folder(
-                account,
-                parse_args(args)?,
-            ))
-            .await?)?,
-            ToolName::DeleteFolder => ser(Box::pin(folder_management::handle_delete_folder(
-                account,
-                parse_args(args)?,
-            ))
-            .await?)?,
-            ToolName::AddLabel => {
-                ser(Box::pin(labels::handle_add_label(account, parse_args(args)?)).await?)?
-            }
-            ToolName::RemoveLabel => {
-                ser(Box::pin(labels::handle_remove_label(account, parse_args(args)?)).await?)?
-            }
-            ToolName::ListLabels => {
-                ser(Box::pin(labels::handle_list_labels(account, parse_args(args)?)).await?)?
-            }
-            ToolName::UseAccount | ToolName::ListAccounts => {
-                return Err(rimap_core::RimapError::Internal(
-                    "infrastructure tools must not reach dispatch_tool".into(),
-                ));
-            }
-        };
-        Ok(resp)
-    }
-
-    /// Handle infrastructure tools that bypass account resolution.
-    ///
-    /// Infrastructure tools are not scoped to an account, so their audit
-    /// records record `account: None` regardless of deployment mode.
-    async fn dispatch_infrastructure(
-        &self,
-        tool: ToolName,
-        args: &serde_json::Map<String, serde_json::Value>,
-    ) -> Result<CallToolResult, ErrorData> {
-        // Infrastructure tools have no per-account posture; record an
-        // explicit sentinel so log readers can distinguish these from
-        // per-account dispatches.
-        self.run_with_audit_envelope(tool, None, PostureContext::Infrastructure, args, async {
-            // Infrastructure tools bypass posture + breaker, but still
-            // enforce a process-wide rate limit.
-            self.registry.check_infrastructure_rate()?;
-            match tool {
-                ToolName::UseAccount => {
-                    let input = parse_args::<crate::tools::admin::accounts::UseAccountInput>(args)?;
-                    ser(
-                        crate::tools::admin::accounts::handle_use_account(&self.registry, input)
-                            .await?,
-                    )
-                }
-                ToolName::ListAccounts => {
-                    ser(crate::tools::admin::accounts::handle_list_accounts(&self.registry).await?)
-                }
-                ToolName::ListFolders
-                | ToolName::Search
-                | ToolName::SearchAdvanced
-                | ToolName::FetchMessage
-                | ToolName::FetchMessageHtml
-                | ToolName::ListAttachments
-                | ToolName::DownloadAttachment
-                | ToolName::MarkRead
-                | ToolName::MarkUnread
-                | ToolName::Flag
-                | ToolName::Unflag
-                | ToolName::AddLabel
-                | ToolName::RemoveLabel
-                | ToolName::ListLabels
-                | ToolName::MoveMessage
-                | ToolName::CreateDraft
-                | ToolName::SendEmail
-                | ToolName::DeleteMessage
-                | ToolName::Expunge
-                | ToolName::CreateFolder
-                | ToolName::RenameFolder
-                | ToolName::DeleteFolder => Err(rimap_core::RimapError::Internal(format!(
-                    "not an infrastructure tool: {}",
-                    tool.as_str(),
-                ))),
-            }
-        })
-        .await
-    }
-
-    /// Apply the registered [`rimap_audit::redact::Redactor`] schema for
-    /// `tool`. If no schema matches, returns an empty object and emits
-    /// a `warn!` — the schema registry is expected to cover every
-    /// advertised tool.
-    fn redact_tool_args(&self, tool: ToolName, args: &serde_json::Value) -> serde_json::Value {
-        if let Some(schema) = self.redaction_schemas.get(&tool) {
-            Redactor::new(schema, self.redaction_salt.as_ref()).apply(args)
-        } else {
-            tracing::warn!(
-                tool = tool.as_str(),
-                "no redaction schema for tool; recording empty arguments_redacted",
-            );
-            serde_json::Value::Object(serde_json::Map::new())
-        }
-    }
-
-    /// Emit a `tool_start` audit record via `spawn_blocking`. Returns the
-    /// allocated `seq` on success; on audit failure emits a `warn!` and
-    /// returns a synthetic `Seq::FIRST` so the call can proceed.
-    ///
-    /// Errors bubble up only when `fail_open = false` AND the write fails:
-    /// in that case the tool call MUST fail because the audit trail is
-    /// broken. `fail_open = true` deployments swallow the error inside
-    /// the writer and return `Ok`.
-    async fn emit_tool_start(
-        &self,
-        tool: ToolName,
-        account: Option<String>,
-        posture: PostureContext,
-        redacted: serde_json::Value,
-        hash: String,
-    ) -> Result<rimap_audit::Seq, ErrorData> {
-        let audit = self.audit.clone();
-        let posture_effective = posture.posture();
-        let join = tokio::task::spawn_blocking(move || {
-            audit.log_tool_start(tool, account.as_deref(), posture_effective, redacted, hash)
-        })
-        .await;
-        match join {
-            Ok(Ok(seq)) => Ok(seq),
-            Ok(Err(audit_err)) => {
-                tracing::error!(error = %audit_err, "tool_start audit write failed");
-                Err(ErrorData::internal_error(
-                    format!("audit write failed: {audit_err}"),
-                    None,
-                ))
-            }
-            Err(join_err) => {
-                tracing::error!(error = %join_err, "tool_start join error");
-                let rimap_err = crate::mcp::spawn_blocking_panic_error(&join_err);
-                Err(crate::mcp::error::to_mcp_error(&rimap_err))
-            }
-        }
-    }
-
-    /// Emit a `tool_end` audit record via `spawn_blocking`. Failures are
-    /// logged but not propagated — at end-of-call the tool has already
-    /// finished and the caller sees its original result.
-    async fn emit_tool_end(
-        &self,
-        start_seq: rimap_audit::Seq,
-        tool: ToolName,
-        account: Option<String>,
-        status: ToolStatus,
-        error_code: Option<rimap_core::ErrorCode>,
-        duration_ms: u64,
-    ) {
-        let audit = self.audit.clone();
-        // The provenance ring buffer is not yet wired for multi-account.
-        // Record an empty snapshot with the window placeholder until a
-        // per-account buffer lands.
-        let provenance = Provenance {
-            window_seconds: 60,
-            message_ids_recently_read: Vec::new(),
-        };
-        let summary = ResultSummary::default();
-        let join = tokio::task::spawn_blocking(move || {
-            audit.log_tool_end(
-                start_seq,
-                tool,
-                account.as_deref(),
-                status,
-                error_code,
-                duration_ms,
-                summary,
-                provenance,
-            )
-        })
-        .await;
-        match join {
-            Ok(Ok(_)) => {}
-            Ok(Err(audit_err)) => {
-                tracing::error!(error = %audit_err, "tool_end audit write failed");
-            }
-            Err(join_err) => {
-                let rimap_err = crate::mcp::spawn_blocking_panic_error(&join_err);
-                tracing::error!(error = %rimap_err, "tool_end join error");
-            }
-        }
-    }
-}
-
-/// Serialize a typed response to `serde_json::Value`.
-///
-/// Used in `dispatch_tool` and `dispatch_infrastructure` to unify
-/// concrete handler return types into a single `Value` before the
-/// audit envelope processes them.
-fn ser<T: serde::Serialize>(resp: T) -> Result<serde_json::Value, rimap_core::RimapError> {
-    serde_json::to_value(&resp).map_err(|e| {
-        rimap_core::RimapError::Internal(format!("response serialization failed: {e}"))
-    })
-}
-
-/// Deserialize tool arguments into a typed input struct.
-fn parse_args<T: serde::de::DeserializeOwned>(
-    args: &serde_json::Map<String, serde_json::Value>,
-) -> Result<T, rimap_core::RimapError> {
-    serde_json::from_value(serde_json::Value::Object(args.clone()))
-        .map_err(|e| rimap_core::RimapError::invalid_input(format!("invalid arguments: {e}")))
-}
-
-/// Convert a `schemars::JsonSchema` type into a JSON object map
-/// suitable for an MCP tool's `inputSchema`.
-fn schema_map<T: schemars::JsonSchema>() -> serde_json::Map<String, serde_json::Value> {
-    let schema = schemars::schema_for!(T);
-    match serde_json::to_value(schema) {
-        Ok(serde_json::Value::Object(mut map)) => {
-            // Strip Rust struct name to avoid leaking implementation
-            // details in the MCP list_tools response.
-            map.remove("title");
-            map
-        }
-        Ok(
-            serde_json::Value::Null
-            | serde_json::Value::Bool(_)
-            | serde_json::Value::Number(_)
-            | serde_json::Value::String(_)
-            | serde_json::Value::Array(_),
-        )
-        | Err(_) => serde_json::Map::new(),
-    }
-}
-
-/// Whether the registry holds exactly one account and its id is the
-/// legacy `"default"` value. Used to preserve bare (non-namespaced)
-/// tool names for single-account deployments.
-fn is_legacy_single_account(
-    accounts: &std::collections::BTreeMap<AccountId, AccountState>,
-) -> bool {
-    accounts.len() == 1
-        && accounts
-            .keys()
-            .next()
-            .is_some_and(|id| id.as_str() == DEFAULT_ACCOUNT_NAME)
-}
-
-/// Split a possibly-namespaced MCP tool name into `(account, tool)`.
-///
-/// Preserves sub-capability tool names that contain dots (e.g.
-/// Promote a base `ToolName` to a sub-capability variant based on args.
-/// Keeps sub-capability posture checks at the dispatch seam rather than
-/// scattered across handlers.
-fn refine_tool_name(
-    base: ToolName,
-    args: Option<&serde_json::Map<String, serde_json::Value>>,
-) -> ToolName {
-    let Some(args) = args else {
-        return base;
-    };
-    // Exhaustive match so a new ToolName variant forces an explicit
-    // refinement decision at the dispatch seam rather than silently
-    // falling through a catch-all.
-    match base {
-        ToolName::FetchMessage
-            if args
-                .get("include_html")
-                .and_then(serde_json::Value::as_bool)
-                == Some(true) =>
-        {
-            ToolName::FetchMessageHtml
-        }
-        ToolName::Search if args.get("advanced_query").is_some() => ToolName::SearchAdvanced,
-        ToolName::ListFolders
-        | ToolName::Search
-        | ToolName::SearchAdvanced
-        | ToolName::FetchMessage
-        | ToolName::FetchMessageHtml
-        | ToolName::ListAttachments
-        | ToolName::DownloadAttachment
-        | ToolName::MarkRead
-        | ToolName::MarkUnread
-        | ToolName::Flag
-        | ToolName::Unflag
-        | ToolName::AddLabel
-        | ToolName::RemoveLabel
-        | ToolName::ListLabels
-        | ToolName::MoveMessage
-        | ToolName::CreateDraft
-        | ToolName::SendEmail
-        | ToolName::DeleteMessage
-        | ToolName::Expunge
-        | ToolName::CreateFolder
-        | ToolName::RenameFolder
-        | ToolName::DeleteFolder
-        | ToolName::UseAccount
-        | ToolName::ListAccounts => base,
-    }
-}
-
-/// `search.advanced_query`): if the raw name parses as a `ToolName`
-/// directly, return it as bare.
-fn split_tool_name(raw: &str) -> (Option<&str>, &str) {
-    if ToolName::from_str(raw).is_ok() {
-        return (None, raw);
-    }
-    match raw.split_once('.') {
-        Some((prefix, rest))
-            if is_valid_account_prefix(prefix) && ToolName::from_str(rest).is_ok() =>
-        {
-            (Some(prefix), rest)
-        }
-        Some(_) | None => (None, raw),
-    }
-}
-
-/// Structural check on an account-namespace prefix. Mirrors the
-/// `AccountId` character rules (ASCII alphanumerics + hyphens, 1–64
-/// chars) without allocating.
-fn is_valid_account_prefix(s: &str) -> bool {
-    !s.is_empty() && s.len() <= 64 && s.chars().all(|c| c.is_ascii_alphanumeric() || c == '-')
-}
-
-/// Type alias for tool spec tuples — `(description, schema)`. The wire
-/// name comes from `ToolName::as_str()` so there is a single source of
-/// truth for tool names.
-type ToolSpec = (&'static str, serde_json::Map<String, serde_json::Value>);
-
-/// Return (description, schema) for the given `ToolName`, or `None`
-/// for sub-capabilities that share an MCP tool name with a parent
-/// (e.g. `SearchAdvanced`, `FetchMessageHtml`).
-fn tool_spec(name: ToolName) -> Option<ToolSpec> {
-    use crate::tools::admin::accounts::UseAccountInput;
-    use crate::tools::compose::create_draft::CreateDraftInput;
-    use crate::tools::compose::send_email::SendEmailInput;
-    use crate::tools::mailbox::delete_message::DeleteMessageInput;
-    use crate::tools::mailbox::expunge::ExpungeInput;
-    use crate::tools::mailbox::flags::FlagInput;
-    use crate::tools::mailbox::folder_management::{
-        CreateFolderInput, DeleteFolderInput, RenameFolderInput,
-    };
-    use crate::tools::mailbox::labels::{LabelInput, ListLabelsInput};
-    use crate::tools::mailbox::move_message::MoveMessageInput;
-    use crate::tools::retrieval::download_attachment::DownloadAttachmentInput;
-    use crate::tools::retrieval::fetch_message::FetchMessageInput;
-    use crate::tools::retrieval::list_attachments::ListAttachmentsInput;
-    use crate::tools::retrieval::search::SearchInput;
-    let tuple = match name {
-        ToolName::ListFolders => ("List all IMAP folders", serde_json::Map::new()),
-        ToolName::Search => (
-            "Search messages with structured query",
-            schema_map::<SearchInput>(),
-        ),
-        ToolName::FetchMessage => (
-            "Fetch message metadata and text body",
-            schema_map::<FetchMessageInput>(),
-        ),
-        ToolName::ListAttachments => (
-            "List attachments on a message",
-            schema_map::<ListAttachmentsInput>(),
-        ),
-        ToolName::DownloadAttachment => (
-            "Download an attachment to the sandbox directory",
-            schema_map::<DownloadAttachmentInput>(),
-        ),
-        ToolName::MarkRead => ("Mark messages as read", schema_map::<FlagInput>()),
-        ToolName::MarkUnread => ("Mark messages as unread", schema_map::<FlagInput>()),
-        ToolName::Flag => (
-            "Add the flagged flag to messages",
-            schema_map::<FlagInput>(),
-        ),
-        ToolName::Unflag => (
-            "Remove the flagged flag from messages",
-            schema_map::<FlagInput>(),
-        ),
-        ToolName::MoveMessage => (
-            "Move messages to another folder",
-            schema_map::<MoveMessageInput>(),
-        ),
-        ToolName::CreateDraft => (
-            "Create a draft email with $PendingReview flag",
-            schema_map::<CreateDraftInput>(),
-        ),
-        ToolName::SendEmail => ("Send an email via SMTP", schema_map::<SendEmailInput>()),
-        ToolName::DeleteMessage => (
-            "Delete a message (move to Trash)",
-            schema_map::<DeleteMessageInput>(),
-        ),
-        ToolName::Expunge => (
-            "Permanently remove deleted messages from a folder",
-            schema_map::<ExpungeInput>(),
-        ),
-        ToolName::CreateFolder => (
-            "Create a new IMAP folder",
-            schema_map::<CreateFolderInput>(),
-        ),
-        ToolName::RenameFolder => ("Rename an IMAP folder", schema_map::<RenameFolderInput>()),
-        ToolName::DeleteFolder => (
-            "Delete an IMAP folder and all its contents",
-            schema_map::<DeleteFolderInput>(),
-        ),
-        ToolName::AddLabel => (
-            "Add a keyword label to messages",
-            schema_map::<LabelInput>(),
-        ),
-        ToolName::RemoveLabel => (
-            "Remove a keyword label from messages",
-            schema_map::<LabelInput>(),
-        ),
-        ToolName::ListLabels => (
-            "List keyword labels on a message",
-            schema_map::<ListLabelsInput>(),
-        ),
-        ToolName::UseAccount => (
-            "Set the active account for subsequent tool calls",
-            schema_map::<UseAccountInput>(),
-        ),
-        ToolName::ListAccounts => ("List all configured email accounts", serde_json::Map::new()),
-        // Sub-capabilities that share an MCP tool name with a parent
-        // (e.g. `SearchAdvanced` shares `search`; `FetchMessageHtml`
-        // shares `fetch_message`) are advertised under the parent entry,
-        // so they have no standalone spec.
-        ToolName::SearchAdvanced | ToolName::FetchMessageHtml => return None,
-    };
-    Some(tuple)
-}
-
-/// Memoized MCP tool definitions. Built once at first access; each
-/// `list_tools` call reuses the same `Arc<JsonObject>` for schemas.
-static TOOL_DEFS: std::sync::LazyLock<std::collections::HashMap<ToolName, Tool>> =
-    std::sync::LazyLock::new(|| {
-        let mut map = std::collections::HashMap::new();
-        for tn in ToolName::all() {
-            let Some((description, schema)) = tool_spec(tn) else {
-                continue;
-            };
-            map.insert(tn, Tool::new(tn.as_str(), description, Arc::new(schema)));
-        }
-        map
-    });
-
-#[cfg(test)]
-#[expect(clippy::expect_used, reason = "tests")]
-mod tests {
-    use rimap_core::tool::ToolName;
-
-    use super::{TOOL_DEFS, refine_tool_name, rimap_error_to_breaker_reason, split_tool_name};
-
-    #[test]
-    fn breaker_reason_maps_service_failures() {
-        use rimap_authz::breaker::FailureReason;
-        use rimap_core::{ErrorCode, RimapError};
-        let err = RimapError::Imap {
-            code: ErrorCode::Timeout,
-            message: "x".into(),
-            source: None,
-        };
-        assert_eq!(
-            rimap_error_to_breaker_reason(&err),
-            Some(FailureReason::Timeout),
-        );
-        let auth = RimapError::Imap {
-            code: ErrorCode::Auth,
-            message: "x".into(),
-            source: None,
-        };
-        assert_eq!(
-            rimap_error_to_breaker_reason(&auth),
-            Some(FailureReason::Auth),
-        );
-        let tls = RimapError::Imap {
-            code: ErrorCode::Tls,
-            message: "x".into(),
-            source: None,
-        };
-        assert_eq!(
-            rimap_error_to_breaker_reason(&tls),
-            Some(FailureReason::Tls)
-        );
-    }
-
-    #[test]
-    fn breaker_reason_ignores_user_errors() {
-        use rimap_core::RimapError;
-        assert_eq!(
-            rimap_error_to_breaker_reason(&RimapError::invalid_input("bad")),
-            None,
-        );
-        assert_eq!(
-            rimap_error_to_breaker_reason(&RimapError::Internal("bug".into())),
-            None,
-        );
-    }
-
-    #[test]
-    fn split_tool_name_bare() {
-        assert_eq!(split_tool_name("send_email"), (None, "send_email"));
-    }
-
-    #[test]
-    fn split_tool_name_namespaced() {
-        assert_eq!(
-            split_tool_name("work.send_email"),
-            (Some("work"), "send_email"),
-        );
-    }
-
-    #[test]
-    fn split_tool_name_preserves_dotted_sub_capability() {
-        // `search.advanced_query` is a valid ToolName and must not be
-        // interpreted as account="search", tool="advanced_query".
-        assert_eq!(
-            split_tool_name("search.advanced_query"),
-            (None, "search.advanced_query"),
-        );
-        assert_eq!(
-            split_tool_name("fetch_message.include_html"),
-            (None, "fetch_message.include_html"),
-        );
-    }
-
-    #[test]
-    fn split_tool_name_unknown_returns_bare() {
-        // Unknown names pass through; `from_str` at the caller rejects.
-        assert_eq!(split_tool_name("garbage"), (None, "garbage"));
-        assert_eq!(split_tool_name("work.garbage"), (None, "work.garbage"),);
-    }
-
-    #[test]
-    fn split_tool_name_rejects_invalid_prefix() {
-        // Underscore is not valid in an account prefix.
-        assert_eq!(
-            split_tool_name("bad_name.send_email"),
-            (None, "bad_name.send_email"),
-        );
-    }
-
-    #[test]
-    fn tool_definition_covers_all_mcp_tools() {
-        // Sub-capabilities are surfaced via their parent tool's schema, not
-        // as standalone MCP tools, so they do not appear in `TOOL_DEFS`.
-        const SUB_CAPABILITIES: &[ToolName] =
-            &[ToolName::SearchAdvanced, ToolName::FetchMessageHtml];
-        let expected = ToolName::all().len() - SUB_CAPABILITIES.len();
-        let defs: Vec<_> = ToolName::all()
-            .into_iter()
-            .filter_map(|tn| TOOL_DEFS.get(&tn))
-            .collect();
-        assert_eq!(defs.len(), expected);
-    }
-
-    #[test]
-    fn refine_tool_name_promotes_sub_capabilities() {
-        let mut args = serde_json::Map::new();
-        args.insert("include_html".into(), serde_json::Value::Bool(true));
-        assert_eq!(
-            refine_tool_name(ToolName::FetchMessage, Some(&args)),
-            ToolName::FetchMessageHtml,
-        );
-
-        let mut args = serde_json::Map::new();
-        args.insert(
-            "advanced_query".into(),
-            serde_json::Value::String("FROM x".into()),
-        );
-        assert_eq!(
-            refine_tool_name(ToolName::Search, Some(&args)),
-            ToolName::SearchAdvanced,
-        );
-    }
-
-    #[test]
-    fn refine_tool_name_is_identity_for_all_other_variants() {
-        // Any variant without a refinement rule must pass through
-        // unchanged, including when args are absent or irrelevant. A new
-        // ToolName variant will fail to compile in `refine_tool_name`'s
-        // exhaustive match until a refinement decision is made.
-        let mut args = serde_json::Map::new();
-        args.insert("include_html".into(), serde_json::Value::Bool(true));
-        args.insert(
-            "advanced_query".into(),
-            serde_json::Value::String("FROM x".into()),
-        );
-        for name in ToolName::all() {
-            let refined_no_args = refine_tool_name(name, None);
-            assert_eq!(refined_no_args, name, "{name:?} changed with no args");
-            let refined = refine_tool_name(name, Some(&args));
-            match name {
-                ToolName::FetchMessage => assert_eq!(refined, ToolName::FetchMessageHtml),
-                ToolName::Search => assert_eq!(refined, ToolName::SearchAdvanced),
-                other => assert_eq!(refined, other, "{other:?} unexpectedly refined"),
-            }
-        }
-    }
-
-    #[test]
-    fn sub_capabilities_return_none() {
-        assert!(TOOL_DEFS.get(&ToolName::SearchAdvanced).is_none());
-        assert!(TOOL_DEFS.get(&ToolName::FetchMessageHtml).is_none());
-    }
-
-    #[test]
-    fn tool_names_are_snake_case() {
-        for def in ToolName::all()
-            .into_iter()
-            .filter_map(|tn| TOOL_DEFS.get(&tn))
-        {
-            assert!(
-                def.name.chars().all(|c| c.is_ascii_lowercase() || c == '_'),
-                "tool name {} is not snake_case",
-                def.name,
-            );
-        }
-    }
-
-    #[test]
-    fn tool_definitions_have_non_empty_schemas() {
-        for def in ToolName::all()
-            .into_iter()
-            .filter_map(|tn| TOOL_DEFS.get(&tn))
-        {
-            // list_folders and list_accounts have no input — empty
-            // schema is expected.
-            if def.name == "list_folders" || def.name == "list_accounts" {
-                continue;
-            }
-            let schema = &def.input_schema;
-            assert!(
-                !schema.is_empty(),
-                "tool {} has empty input schema",
-                def.name,
-            );
-        }
-    }
-
-    #[test]
-    fn every_tool_has_a_description() {
-        for def in ToolName::all()
-            .into_iter()
-            .filter_map(|tn| TOOL_DEFS.get(&tn))
-        {
-            assert!(
-                def.description.is_some(),
-                "tool {} missing description",
-                def.name,
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn infrastructure_tool_emits_tool_start_and_tool_end() {
-        use std::collections::BTreeMap;
-
-        use rimap_audit::{AuditOptions, AuditWriter, Seq};
-        use tempfile::TempDir;
-
-        use crate::boot::registry::AccountRegistry;
-        use crate::mcp::server::ImapMcpServer;
-
-        let tmp = TempDir::new().expect("tempdir");
-        let audit_path = tmp.path().join("audit.jsonl");
-        let audit = AuditWriter::open(&AuditOptions {
-            path: audit_path.clone(),
-            rotate_bytes: 0,
-            rotate_keep: 0,
-            retention_seconds: None,
-            fail_open: false,
-            initial_seq: Seq::FIRST,
-        })
-        .expect("audit open");
-
-        let registry = AccountRegistry::new(BTreeMap::new());
-        let server = ImapMcpServer::new(registry, audit, tmp.path().to_path_buf());
-
-        // list_accounts needs no args and no IMAP connection.
-        let args = serde_json::Map::new();
-        let _ = server
-            .dispatch_infrastructure(ToolName::ListAccounts, &args)
-            .await
-            .expect("list_accounts dispatch");
-
-        drop(server);
-
-        let contents = std::fs::read_to_string(&audit_path).expect("read audit log");
-        let records: Vec<serde_json::Value> = contents
-            .lines()
-            .map(|line| serde_json::from_str(line).expect("parse record"))
-            .collect();
-
-        let start = records
-            .iter()
-            .find(|r| r["kind"] == "tool_start" && r["tool"] == "list_accounts")
-            .expect("tool_start record");
-        let end = records
-            .iter()
-            .find(|r| r["kind"] == "tool_end" && r["tool"] == "list_accounts")
-            .expect("tool_end record");
-
-        assert_eq!(start["seq"], end["start_seq"]);
-        assert_eq!(end["status"], "ok");
-        assert!(start["account"].is_null());
-        assert!(end["account"].is_null());
     }
 }
